requirements.txt
streamlit==1.36.0
pandas==2.2.1
matplotlib==3.9.2
fastapi==0.112.0
uvicorn==0.30.5
pydantic==2.6.3
requests==2.32.3
python-dotenv==1.0.1
openai==0.28.0
pytest==8.3.3
transformers==4.30.2
torch==2.0.1